+++ src/utils/settings.py
@@ -374,4 +374,7 @@ PASSWORDS_TXT_FILE = os.path.abspath(os.path.join(os.path.dirname(__file__), '..
 
 REQUIRED_AUTHENTICATION = False
 
+# Supported HTTP Authentication types
+SUPPORTED_HTTP_AUTH_TYPES = [ "basic", "digest" ]
+
 #eof